Разделила utils.js на render и на common
