+++ src/components/task-edit.js
@@ -1,5 +1,5 @@
 import {COLORS, DAYS, MONTH_NAMES} from "../consts";
-import {formatTime} from "../utils";
+import {formatTime} from "../utils/utils";
 import AbstractComponent from "./abstract-component";
 
 const createRepeatingDaysMarkup = (days, repeatingDays) => {
